rename reward_index -> db, break out extract functions
