+++ reward_index/src/db.rs
