+++ reward_index/src/lib.rs
@@ -1,5 +1,6 @@
+pub mod db;
+pub mod extract;
 pub mod indexer;
-mod reward_index;
 pub mod settings;
 pub mod telemetry;
